+++ apps/party/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gtmgrid/party
 
+## 0.21.0
+
+### Patch Changes
+
+- @gtmgrid/auth@0.21.0
+
 ## 0.20.1
 
 ### Patch Changes